+++ CHANGELOG.md
@@ -1,9 +1,29 @@
-# Change Log
+# Change Log : GitHub Repository Manager
 
-All notable changes to the "github-repository-cloner" extension will be documented in this file.
+### 0.2.0 - May 11, 2020
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+Fixed Cloned Repositories Search not working with some remotes.
 
-## [Unreleased]
+Changed project logo.
 
-- Initial release
+Improved README.
+
+- Added usage .gif (that took my time to get it "ok"!)
+
+Added error message if the extension can't open the OAuth callback server.
+
+Changed REST requests to Graphql. Reduces network usage, reduces time to retrieve the data and also now shows the "Fork of" information when hovering a repository.
+
+### 0.1.1 - May 6, 2020
+
+Quick fixes
+
+### 0.1.0 - May 6, 2020
+
+First release
+
+## Donation
+
+Help me to keep and improve this project!
+
+[![](https://www.paypalobjects.com/en_US/i/btn/btn_donateCC_LG.gif)](https://www.paypal.com/cgi-bin/webscr?cmd=_s-xclick&hosted_button_id=6P2HYMMC2VWMG)